Consolidate the duplicated PgnProtocol into Shared/AgOpenWeb.Models
PgnProtocol existed as two hand-maintained copies: one in the standalone
vehicle simulator, one in AgOpenWeb.IntegrationTests' virtual UDP modules.
They were byte-identical apart from the namespace and a Describe() helper
the simulator added, so they had not drifted yet — but nothing was stopping
them, and this is the module-side definition of the wire format the host
side must agree with.

Move the simulator's copy (the superset) to
Shared/AgOpenWeb.Models/Communication/PgnProtocol.cs and delete the other.
AgOpenWeb.Models is the right home: both consuming projects already
reference it, so no csproj changes are needed, and PgnMessage — the
host-side equivalent — already lives there.

Expand the class doc with two things that were previously unwritten. First,
the checksum rule here (sum of bytes [2 .. len-2]) has to stay in lockstep
with PgnBuilder and PgnMessage.CalculateCRC. Second, IsValidPacket is
genuinely enforced: the virtual modules drop packets that fail it without
logging, so a host-side checksum bug surfaces as silence from the simulator
rather than as an error. That is exactly how the AgIO placeholder CRC in
the hello packet stayed hidden.

Pure refactor — no behaviour change. Simulator builds; Services 1026
passed / 2 skipped, Models 146, ViewModels 222.

Leaves the other four duplicated virtual modules alone. VirtualGpsReceiver,
VirtualMachineModule, VirtualModuleHub and VirtualSteerModule have genuinely
diverged between the two trees (45-153 diff lines each), so reconciling them
involves behavioural decisions and is not a move.
