+++ Simulators/AgOpenWeb.VehicleSimulator/Modules/VirtualSteerModule.cs
@@ -11,6 +11,7 @@ using System.Net;
 using System.Net.Sockets;
 using System.Threading;
 using System.Threading.Tasks;
+using AgOpenWeb.Models.Communication;
 
 namespace AgOpenWeb.VehicleSimulator.Modules;
