+++ Tests/AgOpenWeb.IntegrationTests/VirtualModules/VirtualSteerModuleTests.cs
@@ -11,6 +11,7 @@ using System.Collections.Generic;
 using System.Net;
 using System.Net.Sockets;
 using System.Threading;
+using AgOpenWeb.Models.Communication;
 using NUnit.Framework;
 
 namespace AgOpenWeb.IntegrationTests.VirtualModules;
