+++ Tests/AgOpenWeb.Services.Tests/MachineConfigPgnTests.cs
@@ -11,6 +11,7 @@ using System.Net;
 using System.Net.Sockets;
 using System.Threading.Tasks;
 using AgOpenWeb.IntegrationTests.VirtualModules;
+using AgOpenWeb.Models.Communication;
 using AgOpenWeb.Models.Configuration;
 using AgOpenWeb.Services.AutoSteer;
